+++ CHANGES.rst
@@ -4,7 +4,11 @@ Changes
 0.2 (unreleased)
 ----------------
 
-- No changes yet.
+- Fix control layers creation in ``Map.create_controls``.
+  [rnix]
+
+- Implement static map markers.
+  [rnix]
 
 
 0.1 (2021-11-21)